+++ src/dactyl_keyboard/cad/auxf.clj
@@ -123,33 +123,32 @@
 ;; Ports ;;
 ;;;;;;;;;;;
 
-(defn- port-base
+(defn- port-hole-base-shape
   [getopt id [x y z]]
   (if (= (getopt :ports id :type) :custom-cylindroid)
-    (->> (model/cylinder (max x z) y)
-      (model/rotate [(/ π 2) 0 0])
-      (model/resize [x y z]))
+    (model/resize [x y z] (model/cylinder (max x y) z))
     (model/cube x y z)))
 
 (defn port-hole-base
   "Negative space for one port."
   [getopt id]
   (let [[[_ x] [_ y] z] (place/port-hole-size getopt id)]
-    (port-base getopt id [x y z])))
+    (port-hole-base-shape getopt id [x y z])))
 
 (defn- port-hole-flared
-  "This comes with a flared front plate. This version is suitable for use as
+  "A negative shape that is flared on top. This version is suitable for use as
   a port model, for easier entry in case of imperfect alignment, but it is
   not suitable for use in a tweak because convex hull would widen the hole."
   [getopt id]
-  (let [[[_ x] [_ y] z] (place/port-hole-size getopt id)]
+  (let [[[_ x] [_ y] z] (place/port-hole-size getopt id)
+        size (fn [increment] [(+ x increment) (+ y increment) misc/wafer])]
     (model/union
       (port-hole-base getopt id)
-      (model/translate [0 (/ y 2) 0]
+      (model/translate [0 0 (/ z 2)]
         (model/hull
-          (port-base getopt id [x misc/wafer z])
-          (model/translate [0 1 0]
-            (port-base getopt id [(inc x) misc/wafer (inc z)])))))))
+          (port-hole-base-shape getopt id (size 0))
+          (model/translate [0 0 1]
+            (port-hole-base-shape getopt id (size 1))))))))
 
 (defn port-holder
   "Positive space for one port. Take the ID of the port, not the holder."
@@ -159,7 +158,7 @@
   (let [[x y z] (place/port-holder-size getopt id)]
     (maybe/translate
       (place/port-holder-offset getopt {:anchor id})
-      (port-base getopt id [x y z]))))
+      (port-hole-base-shape getopt id [x y z]))))
 
 (defn port-tweak-post
   "A cube the thickness of the wall around a specific holder."